Qualification Quiz  week0
Much of the quiz is about C programming; most of the code in the course
will be presented in C. For the first two weeks, this is critical: the
low-level errors and defenses against them are germane to C and
low-level programming. Students comfortable in other languages may be
able to learn the necessary C concepts on their own. One book I
recommend is K.A. Reek's Pointers on C. Several previous students of
this course recommended the free book, Learning C the Hard Way.